refactor: change directories name of SPA structure
