+++ src/index.html
@@ -11,7 +11,7 @@
       rel="stylesheet"
     />
     <title>Pet Hat</title>
-    <script src="./pages/index.js" type="module"></script>
+    <script src="./layouts/index.js" type="module"></script>
     <link
       rel="icon"
       type="image/x-icon"
