+++ src/layouts/pages/NoPetRegirestedPage/index.scss
