+++ src/layouts/pages/PetRegisterPage/index.scss
